Change buffer lengths, offsets, etc from 'unsigned int' to 'size_t'

+++ src/arch.h
@@ -37,9 +37,9 @@ typedef void (*_aax_free_proc)(void*);
 
 typedef void (*_batch_cvt_proc)(void*, size_t);
 typedef void (*_batch_cvt_from_proc)(void_ptr, const_void_ptr, size_t);
-typedef void (*_batch_cvt_from_intl_proc)(int32_ptrptr, const_void_ptr, size_t, size_t, size_t);
+typedef void (*_batch_cvt_from_intl_proc)(int32_ptrptr, const_void_ptr, size_t, unsigned int, size_t);
 typedef void (*_batch_cvt_to_proc)(void_ptr, const_void_ptr, size_t);
-typedef void (*_batch_cvt_to_intl_proc)(void_ptr, const_int32_ptrptr, size_t, size_t, size_t);
+typedef void (*_batch_cvt_to_intl_proc)(void_ptr, const_int32_ptrptr, size_t, unsigned int, size_t);
 typedef void (*_batch_codec_proc)(void_ptr, const_void_ptr, unsigned char, size_t);
