+++ src/software/cpu/rbuf2d_effects.h
@@ -25,11 +25,11 @@ extern "C" {
 
 #include "software/rbuf_int.h"
 
-void _aaxRingBufferFilterFrequency(_aaxRingBufferSample*, MIX_PTR_T, CONST_MIX_PTR_T, unsigned int, unsigned int, unsigned int, unsigned int, void*, void *, unsigned char);
-void _aaxRingBufferEffectDistort(_aaxRingBufferSample*, MIX_PTR_T, CONST_MIX_PTR_T, unsigned int, unsigned int, unsigned int, unsigned int, void*, void*);
-void _aaxRingBufferEffectDelay(_aaxRingBufferSample*, MIX_PTR_T, CONST_MIX_PTR_T, MIX_PTR_T, unsigned int, unsigned int, unsigned int, unsigned int, void*, void*, unsigned int);
-void _aaxRingBufferEffectReflections(_aaxRingBufferSample*, MIX_PTR_T, CONST_MIX_PTR_T, MIX_PTR_T, unsigned int, unsigned int, unsigned int, unsigned int, const void*);
-void _aaxRingBufferEffectReverb(_aaxRingBufferSample*, MIX_PTR_T, unsigned int, unsigned int, unsigned int, unsigned int, const void*);
+void _aaxRingBufferFilterFrequency(_aaxRingBufferSample*, MIX_PTR_T, CONST_MIX_PTR_T, size_t, size_t, size_t, unsigned int, void*, void *, unsigned char);
+void _aaxRingBufferEffectDistort(_aaxRingBufferSample*, MIX_PTR_T, CONST_MIX_PTR_T, size_t, size_t, size_t, unsigned int, void*, void*);
+void _aaxRingBufferEffectDelay(_aaxRingBufferSample*, MIX_PTR_T, CONST_MIX_PTR_T, MIX_PTR_T, size_t, size_t, size_t, size_t, void*, void*, unsigned int);
+void _aaxRingBufferEffectReflections(_aaxRingBufferSample*, MIX_PTR_T, CONST_MIX_PTR_T, MIX_PTR_T, size_t, size_t, size_t, unsigned int, const void*);
+void _aaxRingBufferEffectReverb(_aaxRingBufferSample*, MIX_PTR_T, size_t, size_t, size_t, unsigned int, const void*);
 
 #if defined(__cplusplus)
 }  /* extern "C" */
